Функции render и unrender из файла utils.js сделаны методами класса PageController
